requirements.txt
openai
streamlit
pandas
azure-identity
